Use built-in GetHashCode and change tickCount to ulong throughout

+++ backend/spacetimedb/GameAI/GameAIBehavior.cs
@@ -7,7 +7,7 @@ using static Module;
 
 public static partial class GameAI
 {
-    public static Tank EvaluateAndMutateTank(ReducerContext ctx, FullTank fullTank, GameAIContext aiContext, int tickCount)
+    public static Tank EvaluateAndMutateTank(ReducerContext ctx, FullTank fullTank, GameAIContext aiContext, ulong tickCount)
     {
         var tank = fullTank.Tank;
         bool canFireThisTick = BehaviorTreeAI.CanBotFireOnTick(tank.Id, tickCount);
